fixed GCC compile warnings

+++ kernels/subdiv/feature_adaptive_eval_grid.h
@@ -202,7 +202,7 @@ namespace embree
       
       __forceinline bool final(const CatmullClarkPatch3fa& patch, const CatmullClarkRing::Type type, size_t depth) 
       {
-        const int max_eval_depth = (type & CatmullClarkRing::TYPE_CREASES) ? PATCH_MAX_EVAL_DEPTH_CREASE : PATCH_MAX_EVAL_DEPTH_IRREGULAR;
+        const size_t max_eval_depth = (type & CatmullClarkRing::TYPE_CREASES) ? PATCH_MAX_EVAL_DEPTH_CREASE : PATCH_MAX_EVAL_DEPTH_IRREGULAR;
 //#if PATCH_MIN_RESOLUTION
 //        return patch.isFinalResolution(PATCH_MIN_RESOLUTION) || depth>=max_eval_depth;
 //#else
